requirements.txt
ultralytics>=8.3.0
streamlit>=1.32.0
Pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0
